better method for highlighting a cell widget.
instead of using setPalette for the cell widget, we paint a a filled rectangle in paintEvent and depending on the variable self.frameHighlight, this rectangle is drawn with a different color.
Also the cell widget that has focus is drawn with a red innter Frame. And if the corresponding cell is solved, the number is drawn red as well. The Color is saved in the hash SudokuCellWidget.Colors and can be changed easily.