build the background higlighting into the main program
this was done for the elimination of values, based on solved cells.
in sudoku_cell_widget.py: wrote function setHaloValues to make
a cell display halo background highlighting.
in sudoku_board_widget.py: wrote functions _setIndicesHalo,
setRowHalo, setColHalo, setSquareHalo to display this halo for a whole unit
for a certain set of values.

Also improved the _row_indices function in board.py